urlService/requirements.txt
Django~=5.0.6
django-cors-headers==4.3.1
setuptools==57.0.0
djangorestframework==3.15.1
uwsgi